+++ inc/ti/pkg.h
@@ -19,11 +19,8 @@ void ti_pkg_log(ti_pkg_t * pkg);
 void ti_pkg_set_tp(ti_pkg_t * pkg, uint8_t tp);
 static inline size_t ti_pkg_sz(ti_pkg_t * pkg);
 
-
-/* setting ntp is to avoid ~ unsigned warn */
 #define ti_pkg_check(pkg__) (\
-        (pkg__)->tp == ((pkg__)->ntp ^= 0xff) && \
-        (pkg__)->tp != ((pkg__)->ntp ^= 0xff) && \
+        (pkg__)->tp == ((pkg__)->ntp ^ 0xff) && \
         (pkg__)->n <= TI_PKG_MAX_SIZE)
 
 /* return total package size, header + data size */