+++ itest/test_gc.py
@@ -4,9 +4,6 @@ from lib import run_test
 from lib import default_test_setup
 from lib.testbase import TestBase
 from lib.client import get_client
-from thingsdb.exceptions import AuthError
-from thingsdb.exceptions import ForbiddenError
-from thingsdb.exceptions import BadDataError
 
 
 class TestGC(TestBase):