+++ src/ti/backup.c
@@ -276,7 +276,7 @@ char * ti_backup_gcloud_task(ti_backup_t * backup)
     return buf.data;
 }
 
-char * ti_backup_task(ti_backup_t * backup)
+char * ti_backup_file_task(ti_backup_t * backup)
 {
     struct tm * tm_info;
     uint64_t now = util_now_usec();