+++ src/ti/ctask.c
@@ -409,6 +409,187 @@ static int ctask__new_type(ti_thing_t * thing, mp_unp_t * up)
 }
 
 /*
+ * The new_enum MUST be followed with set_enum_data to set the actual members
+ * and methods. This creates place-holders for members to they can be used
+ * while unpacking data.
+ */
+static int ctask__new_enum(ti_thing_t * thing, mp_unp_t * up)
+{
+    ex_t e = {0};
+    ti_collection_t * collection = thing->collection;
+    ti_enum_t * enum_;
+    mp_obj_t obj, mp_id, mp_name, mp_created, mp_size;
+
+    if (mp_next(up, &obj) != MP_MAP || obj.via.sz != 4 ||
+        mp_skip(up) != MP_STR ||
+        mp_next(up, &mp_id) != MP_U64 ||
+        mp_skip(up) != MP_STR ||
+        mp_next(up, &mp_created) != MP_U64 ||
+        mp_skip(up) != MP_STR ||
+        mp_next(up, &mp_name) != MP_STR ||
+        mp_skip(up) != MP_STR ||
+        mp_next(up, &mp_size) != MP_U64)
+    {
+        log_critical(
+            "task `new_enum` for "TI_COLLECTION_ID" is invalid",
+            collection->id);
+        return -1;
+    }
+    enum_ = ti_enums_by_id(collection->enums, mp_id.via.u64);
+    if (enum_)
+    {
+        log_critical(
+                "task `new_enum` for "TI_COLLECTION_ID" is invalid; "
+                "enum with id %"PRIu64" already found",
+                collection->id, mp_id.via.u64);
+        return -1;
+    }
+    enum_ = ti_enum_create(
+            mp_id.via.u64,
+            mp_name.via.str.data,
+            mp_name.via.str.n,
+            mp_created.via.u64,
+            0);  /* modified_at = 0 */
+    if (!enum_ || ti_enums_add(collection->enums, enum_))
+    {
+        log_critical(EX_MEMORY_S);
+        goto fail0;
+    }
+
+    if (ti_enum_create_placeholders(enum_, mp_size.via.u64, &e))
+    {
+        log_critical(e.msg);
+        goto fail1;
+    }
+
+    return 0;
+fail1:
+    ti_enums_del(collection->enums, enum_, NULL);
+fail0:
+    ti_enum_destroy(enum_);
+    return -1;
+}
+
+static int ctask__set_enum_data(ti_thing_t * thing, mp_unp_t * up)
+{
+    ex_t e = {0};
+    ti_collection_t * collection = thing->collection;
+    ti_enum_t * enum_;
+    mp_obj_t obj, mp_id;
+    ti_vup_t vup = {
+            .isclient = false,
+            .collection = collection,
+            .up = up,
+    };
+
+    if (mp_next(up, &obj) != MP_MAP || obj.via.sz != 3 ||
+        mp_skip(up) != MP_STR ||
+        mp_next(up, &mp_id) != MP_U64 ||
+        mp_skip(up) != MP_STR)
+    {
+        log_critical(
+            "task `set_enum_data` for "TI_COLLECTION_ID" is invalid",
+            collection->id);
+        return -1;
+    }
+
+    enum_ = ti_enums_by_id(collection->enums, mp_id.via.u64);
+    if (!enum_)
+    {
+        log_critical(
+                "task `set_enum_data` for "TI_COLLECTION_ID" is invalid; "
+                "enum with id %"PRIu64" not found",
+                collection->id, mp_id.via.u64);
+        return -1;
+    }
+
+    if (ti_enum_set_members_from_vup(enum_, &vup, &e))
+    {
+        log_critical(e.msg);
+        return -1;
+    }
+
+    if (ti_enum_init_methods_from_vup(enum_, &vup, &e))
+    {
+        log_critical(e.msg);
+        return -1;
+    }
+
+    return 0;
+}
+
+static int ctask__import(ti_thing_t * thing, mp_unp_t * up)
+{
+    int rc = 0;
+    ex_t e = {0};
+    mp_obj_t obj, mp_import_tasks;
+    ti_collection_t * collection = thing->collection;
+
+    if (mp_next(up, &obj) != MP_MAP || obj.via.sz != 2 ||
+        mp_skip(up) != MP_STR ||
+        mp_next(up, &mp_import_tasks) != MP_BOOL ||
+        mp_skip(up) != MP_STR)
+    {
+        log_critical(
+            "task `import` for "TI_COLLECTION_ID" is invalid",
+            collection->id);
+        return -1;
+    }
+
+    if (ti_collection_unpack(collection, up, &e))
+    {
+        log_error(e.msg);  /* not critical for sure */
+        rc = -1;
+    }
+
+    if (!mp_import_tasks.via.bool_)
+        ti_collection_tasks_clear(collection);
+
+    return rc;
+}
+
+static int ctask__replace_root(ti_thing_t * thing, mp_unp_t * up)
+{
+    _Bool changed;
+    ti_val_t * val;
+    ti_collection_t * collection = thing->collection;
+    ti_vup_t vup = {
+            .isclient = false,
+            .collection = collection,
+            .up = up,
+    };
+
+    /* drop the collection root thing; this must be removed as the packed
+     * value might contain the same Id the the old collection root; when this
+     * is applied as a change, we must prevent keeping the dropped thing; */
+    changed = ti_changes_unkeep_dropped();
+    ti_val_unsafe_drop((ti_val_t *) collection->root);
+    if (changed)
+        ti_changes_keep_dropped();
+
+    collection->root = NULL;
+
+    /* collection must be empty at this point */
+    assert(collection->things->n == 0);
+
+    val = ti_val_from_vup(&vup);
+    if (!val)
+        return -1;  /* logging is done */
+
+    if (!ti_val_is_thing(val))
+    {
+        log_critical(
+                "task `replace_root` for "TI_COLLECTION_ID" is invalid; "
+                "value is not a thing",
+                collection->id);
+        return -1;
+    }
+
+    collection->root = (ti_thing_t *) val;  /* no reference required */
+    return 0;
+}
+
+/*
  * Returns 0 on success
  * - for example: {'enum_id':.., 'members':.. }
  *
@@ -451,18 +632,20 @@ static int ctask__set_enum(ti_thing_t * thing, mp_unp_t * up)
                 collection->id, mp_id.via.u64);
         return -1;
     }
-
-    enum_ = ti_enum_create(
-            mp_id.via.u64,
-            mp_name.via.str.data,
-            mp_name.via.str.n,
-            mp_created.via.u64,
-            0);  /* modified_at = 0 */
-
-    if (!enum_ || ti_enums_add(collection->enums, enum_))
+    else
     {
-        log_critical(EX_MEMORY_S);
-        goto fail0;
+        enum_ = ti_enum_create(
+                mp_id.via.u64,
+                mp_name.via.str.data,
+                mp_name.via.str.n,
+                mp_created.via.u64,
+                0);  /* modified_at = 0 */
+
+        if (!enum_ || ti_enums_add(collection->enums, enum_))
+        {
+            log_critical(EX_MEMORY_S);
+            goto fail0;
+        }
     }
 
     if (ti_enum_init_members_from_vup(enum_, &vup, &e))
@@ -476,9 +659,6 @@ static int ctask__set_enum(ti_thing_t * thing, mp_unp_t * up)
         log_critical(e.msg);
         goto fail1;
     }
-
-    /* update created time-stamp */
-    enum_->created_at = mp_created.via.u64;
 
     return 0;
 
@@ -2163,6 +2343,9 @@ static int ctask__vtask_new(ti_thing_t * thing, mp_unp_t * up)
     }
 
     user = ti_users_get_by_id(mp_user_id.via.u64);
+    if (!user)
+        user = vec_first(ti.users);
+
     closure = (ti_closure_t *) ti_val_from_vup(&vup);
 
     if (!closure || !ti_val_is_closure((ti_val_t *) closure) ||
@@ -3099,6 +3282,10 @@ int ti_ctask_run(ti_thing_t * thing, mp_unp_t * up)
     case TI_TASK_REN:               return ctask__ren(thing, up);
     case TI_TASK_FILL:              return ctask__fill(thing, up);
     case TI_TASK_MOD_PROCEDURE:     return ctask__mod_procedure(thing, up);
+    case TI_TASK_NEW_ENUM:          return ctask__new_enum(thing, up);
+    case TI_TASK_SET_ENUM_DATA:     return ctask__set_enum_data(thing, up);
+    case TI_TASK_REPLACE_ROOT:      return ctask__replace_root(thing, up);
+    case TI_TASK_IMPORT:            return ctask__import(thing, up);
     }
 
     log_critical("unknown collection task: %"PRIu64, mp_task.via.u64);